philo middle

+++ 42course/philosophers/philo/philo.h
@@ -6,7 +6,7 @@
 /*   By: youngski <youngski@student.42seoul.>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/23 18:23:33 by youngski          #+#    #+#             */
-/*   Updated: 2023/02/24 16:02:45 by youngski         ###   ########.fr       */
+/*   Updated: 2023/02/25 12:27:47 by youngski         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ typedef struct s_arg
 	pthread_mutex_t	print;
 	pthread_mutex_t	*forks;
 	int				philo_num;
-	int				time_to_die;
+	unsigned long	time_to_die;
 	int				time_to_eat;
 	int				time_to_sleep;
 	int				time_to_end;
@@ -39,13 +39,13 @@ typedef struct s_arg
 
 typedef struct	s_philo
 {
-	t_arg		*arg;
-	int			id;
-	int			left;
-	int			right;
-	int			last_eat_time;
-	int			eat_count;
-	pthread_t	thread;
+	t_arg			*arg;
+	int				id;
+	int				left;
+	int				right;
+	unsigned long	last_eat_time;
+	int				eat_count;
+	pthread_t		thread;
 }	t_philo;
 
 void		ft_pass_time(long long wait_time, t_arg *arg);
@@ -57,7 +57,7 @@ void		ft_philo_check_finish(t_arg *arg, t_philo *philo);
 int			ft_arg_init_mutex(t_arg *arg);
 int			ft_arg_init(t_arg *arg, int argc, char *argv[]);
 int			ft_philo_init(t_philo **philo, t_arg *arg);
-long long	ft_get_time(void);
+unsigned long	ft_get_time(void);
 void		ft_free_thread(t_arg *arg, t_philo *philo);
 
 int		print_error(char *msg, int num);
